Move parsing logic into a dedicated namespace caliopen_mail.parsers
